feat(#103): Add EoData object in order to save data into XMIR correctly, as a string of hex bytes
